+++ python/libneo/__init__.py
@@ -15,10 +15,10 @@ from .convert_neo_in_to_fortran_namelist import *
 from .getHeaderDataVMEC import *
 from .flux_converter import *
 from .interpolate import *
-from .SemiPeriodicFourierSpline import SemiPeriodicFourierSpline
-from .SemiPeriodicFourierSpline import SemiPeriodicFourierSpline_ExceedingNyquistLimit
+from .semi_periodic_fourier_spline import SemiPeriodicFourierSpline
+from .semi_periodic_fourier_spline import SemiPeriodicFourierSpline_ExceedingNyquistLimit
 from .fourier_utils import fourier_coefs_half, fourier_coefs_full
 from .fourier_utils import get_half_fft
 
-from .CoordinateConverter import StorGeom2MarsCoords
-from .CoordinateConverter import order_monotonically
+from .coordinate_converter import StorGeom2MarsCoords
+from .coordinate_converter import order_monotonically
