+++ python/libneo/semi_periodic_fourier_spline.py
