+++ tests/python/test_semi_periodic_fourier_spline.py
